http proxy components splited across several files for easier management
