+++ proxy/http_executor.go
@@ -1,0 +1,22 @@
+package proxy
+
+import (
+	"context"
+	"net/http"
+)
+
+// HTTPRequestExecutor defines the interface of the request executor for the HTTP transport protocol
+type HTTPRequestExecutor func(ctx context.Context, req *http.Request) (*http.Response, error)
+
+// DefaultHTTPRequestExecutor creates a HTTPRequestExecutor with the received HTTPClientFactory
+func DefaultHTTPRequestExecutor(clientFactory HTTPClientFactory) HTTPRequestExecutor {
+	return func(ctx context.Context, req *http.Request) (*http.Response, error) {
+		return clientFactory(ctx).Do(req.WithContext(ctx))
+	}
+}
+
+// HTTPClientFactory creates http clients based with the received context
+type HTTPClientFactory func(ctx context.Context) *http.Client
+
+// NewHTTPClient just creates a http default client
+func NewHTTPClient(_ context.Context) *http.Client { return http.DefaultClient }
