+++ proxy/http_response.go
@@ -1,0 +1,43 @@
+package proxy
+
+import (
+	"context"
+	"io"
+	"net/http"
+
+	"github.com/devopsfaith/krakend/encoding"
+)
+
+// HTTPResponseParser defines how the response is parsed from http.Response to Response object
+type HTTPResponseParser func(context.Context, *http.Response) (*Response, error)
+
+// DefaultHTTPResponseParserConfig defines a default HTTPResponseParserConfig
+var DefaultHTTPResponseParserConfig = HTTPResponseParserConfig{
+	func(_ io.Reader, _ *map[string]interface{}) error { return nil },
+	EntityFormatterFunc(func(r Response) Response { return r }),
+}
+
+// HTTPResponseParserConfig contains the config for a given HttpResponseParser
+type HTTPResponseParserConfig struct {
+	Decoder         encoding.Decoder
+	EntityFormatter EntityFormatter
+}
+
+// DefaultHTTPResponseParserFactory creates HTTPResponseParser from a given HTTPResponseParserConfig
+type HTTPResponseParserFactory func(HTTPResponseParserConfig) HTTPResponseParser
+
+// DefaultHTTPResponseParserFactory is the default implementation of HTTPResponseParserFactory
+func DefaultHTTPResponseParserFactory(cfg HTTPResponseParserConfig) HTTPResponseParser {
+	return func(ctx context.Context, resp *http.Response) (*Response, error) {
+		var data map[string]interface{}
+		err := cfg.Decoder(resp.Body, &data)
+		resp.Body.Close()
+		if err != nil {
+			return nil, err
+		}
+
+		newResponse := Response{Data: data, IsComplete: true}
+		newResponse = cfg.EntityFormatter.Format(newResponse)
+		return &newResponse, nil
+	}
+}
